preadv02/pwritev02: fix EFAULT testcase on s390/x
EFAULT testcase is behaving differently on s390, mainly because
access_ok() always returns true, so rw_copy_check_uvector() can't
detect that something is wrong with iovec we pass in.

1. access_ok() on s390/x always return true [1], so we don't
fail at rw_copy_check_uvector() as we do on other arches

  preadv
    vfs_readv
      do_readv_writev
        rw_copy_check_uvector
          access_ok

2. With access_ok() ineffective, do_generic_file_read() returns
failure only if nothing could be read. So we pass just one bad
iovec to trigger EFAULT on s390.

  preadv
    vfs_readv
      do_readv_writev
        xfs_file_read_iter
          generic_file_read_iter
            do_generic_file_read
              written ? written : error;

3. To avoid inode size check in do_generic_file_read() that would
prematurely end this function, we truncate new file to size of 1 page.

[1] d12a2970385c "s390/uaccess: remove pointless access_ok() checks"

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>