fs/isofs.sh: Don't use /etc/ for the image
The content of /etc/ is not predictable and may cause false possitives.

I recently started to get failures with -J when there are two files with
long names that differ only at the end.

The exact error I got is:

genisoimage: Error:
/etc/zypp/repos.d/SUSE_Linux_Enterprise_Server_12_SP1_x86_64:SLES12-SP1-Debuginfo-Updates.repo
and
/etc/zypp/repos.d/SUSE_Linux_Enterprise_Server_12_SP1_x86_64:SLES12-SP1-Debuginfo-Pool.repo
have the same Joliet name
Joliet tree sort failed.

This patch changes the script to generate filesystem tree instead of
copying content of /etc/.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>