syscalls/umount2_02: Fix umount() and EBUSY
We have to use tst_umount() that retries the umount() on EBUSY otherwise
the test will rarely fail to umount the device in the main loop. This
umount is not triggered under normal circumstances, it is however used
when something went wrong to make sure the device is umounted.

The test for umount2() MNT_EXPIRED races with background daemons scans
as well and sometimes the last call to umount2() with MNT_EXPIRED fails
unexpectedly with EBUSY as well (since gvfs-trash has accessed the mount
point meanwhile) which causes the SAFE_UMOUNT() to be called in the
main() which fails with TBROK since the umount() continues to fail with
EBUSY as well and the device is finally umounted in cleanup().

So this is more or less cosmetic change so which makes the test exit
with TFAIL instead of TBROK in this case.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>