syscalls/umount2_{02,03}: Retry umount2() on EBUSY
This should be last fix for umount2() testcases failing with unexpected
EBUSY caused by background daemons probing new mounts (gvfs-trash in
this case).

This patch adds umount2.h header that defined inline umount2_retry()
function that retries umount2() only in case of EBUSY.

The umount2_01 does not need to be fixed since it uses MNT_DETACH that
does lazy umount and hence it does not race with background probing.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>