scenario_groups/default: Remove dma_thread_diotest
Looks like there is no interest in fixing the race between fork() and
direct I/O between the kernel devs [1]. Even the open.2 manual page
states that this is a known problem and that it leads to undefined
behavior. This renders these testcases useless. They fail but that's all
we can do about it.

So let's disable the testcases from default run for now.

We may either try to change them to use madvise(MAD_DONTFORK) or shared
mapping for the buffers which are suggested workarounds, or remove them
completely.

[1] http://www.spinics.net/lists/linux-mm/msg75488.html

The issue was recently brought up in LTP GitHub issue #63.

(It's fairly easy to reproduce with a ext3 fs, for me with 2GB loopback
 all of the dma_thread_diotest have failed every time.)

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>