ftrace_stress: skip unsupported tests and early cleanup
Checking if it's supported before lanuching a test script. This action is
through verifying the matched file in ftrace tracing direcrotry.

At the same time, transform the hard coded pids to the test_targets string and
make it easier to be controlled.

With this modification, the outdated tracing_enabled issue can be skiped on
newer kernels.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>