open_posix_testsuite: Fix parallel build.
The main the Makefile that generates test Makefiles was wrong.

What it did (simplified) was:

CRITICAL_MAKEFILES=conformance functional stress

all: $(CRITICAL_MAKEFILES)

$(CRITICAL_MAKEFILES):
        $(MAKE) generate-makefiles

No wonder that it started $(MAKE) generate-makefiles for each Makefile in
CRITICAL_MAKEFILES and running the generate script concurenlty produced
corrupted Makefiles.

As a matter of fact it started only two instances of the generate-makefiles.sh
script because there was a typo, the CRITICAL_SECTION_MAKEFILE should probably
have been CRITICAL_STRESS_MAKEFILE.

The obvious fix is to change the Makefile to depend only on single critical
Makefile in order to figure out if Makefiles were generated or not.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>