containers/netns/sysfs: check for known iface names, explicitly
The original code tries to check that no existing host interfaces
made it into the new netns. This needs exceptions for ifaces that
are created automatically, ie. 'lo', 'dummy0' and (on newer kernels)
'ip_vti0', which is not a long-term sustainable solution.

It also hade a corner case of always PASSing if the host had no
network interfaces of its own - it only checked for "no extra ifaces
in the new ns".

This change makes it use explicit pre-created interface names in
both the host and the new ns, avoiding any (existing) ones.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>