migrate_pages02: make 'not all pages migrated' a TINFO
Occasionally we get this warning when all references to
a page are not removable at the time of migration (see
Documentation/vm/page_migration), usually a single page
in libc or ld, which was presumably locked into memory.

This creates unnecessary noise in results. Test still checks
that pages allocated by testcase are correctly moved,
so eliminate this sporadic TWARN by turning it into TINFO.

Signed-off-by: Jan Stancek <jstancek@redhat.com>