containers,tomoyo: Fix order of included headers
Since the test.h defined clone(...) macro in order to avoid direct use
of the function it has to be included after all system headers,
otherwise the definition of clone() function in sched.h will be replaced
as well.

This wasn't the case for most of the containers/*/*.c testcases which
included test.h followed by foo_helper.h header that included sched.h.
The code in question compiled by a pure luck as the clone function
definition was replaced by a garbage which had been syntactically
correct.

This patch moves the test.h header after the foo_helper.h and also
removes a few duplicated includes as well.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>