tunable/max_map_count: set a safer lower limit
When run LTP's max_map_count in a GUI terminal, Xorg can easily crashs.
The root cause is Xorg can have hundreds of maps in the normal case. So
we have to set a safer lower limit for the max_map_count file. In this
patch we set it to 1024 which is safe during our observation.

Signed-off-by: Huacai Chen <chenhc@lemote.com>