controllers/memcg: Remove force_empty with locked memory test
The testcase started to fail with newer kernels. The reason for this is
that we only try to reclaim all the memory from the given memcg and
retry MEM_CGROUP_RECLAIM_RETRIES (5) times. Mlocked memory is not
reclaimable so we will simply eat all the retries and return back. We
used to reparent the remaining charges to the parent. This is no longer
true since f61c42a7d911 ("memcg: remove tasks/children test from
mem_cgroup_force_empty()"). The behavior change is documented and
considered complying with the previous documentation.

The main use case for the knob as described by
Documentation/cgroups/memory.txt is to:

"
  The typical use case for this interface is before calling rmdir().
  Because rmdir() moves all pages to parent, some out-of-use page caches can be
  moved to the parent. If you want to avoid that, force_empty will be useful.
"

So this change of behavior was deliberate and considering the testcase is quite
artificial I do not consider this a bug.

Signed-off-by: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>