eject-tests.sh: Rewrite && convert to new library.
The test is not enabled by default though since it needs a CD in the
default CD drive and the drive should be able to close tray as well,
which is not true for some (mostly external USB or notebook ones).

Also the auto-eject feature does not seems to work with my drive, no
idea why.

Maybe we can create a virtual cdrom drive for the test with stgt[1]
but that would require a bit more work.

[1] http://stgt.sourceforge.net/

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>