mem/oom01: Run oom01 test only once
This patch is really not to fix an LTP bug. But if we want to verify
kernel functions, run oom01 only once is enough. Running oom01 for
multiple times may hurt the whole Linux system, because applications
(such as systemd, Xorg, and so on) may have some corner cases that
malloc()/mmap() failures are not handled perfectly.

If we want to verify kernel functions, please run LTP in this manner
and oom01 will run only once:

./runltp
./runltp -t 3d

If we want to discover application bugs, please run LTP in this manner
and oom01 will run multiple times:

while true; do
    ./runltp
done

Signed-off-by: Huacai Chen <chenhc@lemote.com>