syscalls/setpriority01: Use newly created user for the test
The call to setpriority(PRIO_USER, 0, prio) sets the priorities of all
the root processes to the specified value, though this is kind of
expected for the syscall, it impacts the whole system.

So the test now creates a new user and use it for the PRIO_USER tests.
We also fork a child process to be used for the PRIO_PROCESS and
PRIO_PGRP tests in order not to mess with the test environment.

Signed-off-by: Guangwen Feng <fenggw-fnst@cn.fujitsu.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>