du01: Fix the BTRFS fix after test conversion
I've moved the page_size variable inicialization into the setup()
function while converting the testcase to the new library in:

commit d40cbfb2372f68aef72ebdf167018d7374622ded
Author: Cyril Hrubis <chrubis@suse.cz>
Date:   Mon Sep 26 18:31:03 2016 +0200

    du01.sh: Convert to the new library

Which broke the test on BTRFS again, since the page_size has to be
defined before the check3 variable is assigned, otherwise the
${page_size} in the string is evaluated to empty string.

So this commit moves the inicialization back so the page_size variable
is assigned before the check3 variable.

Sorry for the breakage.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>