madvise01: Fix MADV_FREE
Currently MADV_FREE works only on anonymous pages.

Also the test was passing a file descriptor to the MAP_ANONYMOUS mmap(),
which is not wrong since the fd argument is ignored in that case, but it
was mildly confusing. So now we pass -1 and the pointer variable is
called amem instead of pfile which was confusing as well.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>