tst_test: Allow using tst_res and tst_brk without initialized IPC
If we try to print TINFO messages while the shared memory to store
results is not initialized we stop the test execution with TBROK.

But currently there is no way the results not to be allocated when test
code is executed since it's initialized before we call any of the test
functions.

The motivation for this change is that test library code that prints
messages (tst_device.c for instance) cannot be used without doing full
test initialization, which is unfortunate since I wanted to reuse the
code to create a helper binary for shell testcases.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>