lib: Move some code to testcases/kernel/fs/doio/
Quite a number of code in lib/ directory was used only by one of the
three doio testcases. Hence it does not make much sense to keep it in
lib/.

So this commit moves the files + corresponding header files into the
doio directory and also deletes corresponding manual pages (+ few pages
for nonexisting functions as well).

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>