bin/ftest02: Fix negative lseek offset expression.
Cast the result of sizeof to (off_t) so that we have a signed value
before we negate it.  Otherwise we will get the wrong value in cases
where sizeof returns a 32 bit unsigned type and off_t is a 64 bit
signed type.  This problem shows up on x86_64 when compiling with
-mx32.

This fixes LTP issue #135.

Signed-off-by: Steve Ellcey <sellcey@caviumnetworks.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>