syscalls/clock_nanosleep2_01: Rewrite.
The test only calls clock_nanosleep2_01 with absolute timeout. There is
absolutely no reason to do that 50 times with 1s sleep time as the
original test had to be interrupted by gdb in order to trigger the
problem.

So this commit changes the sleep time to 0.1s which saves 50 seconds on
syscall testrun and also converts the test to the new test library.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>