daemonlib: Use is-active instead of status for systemctl
'systemctl status service' exit code is broken in old distros. It
returns 0 even if service is not running. As we use status_daemon for
detecting whether service is running, it's more reliable to use
is-active.

systemctl bug was introduced in systemd v209 (f74294c1d), fixed in v213
(5bb75bc74).

Signed-off-by: Petr Vorel <pvorel@suse.cz>