tst_test.c: set the test id automatically
Use the argv[0] to set the tst_test->tid automatically since we fail to
update it commonly and we end up with wrong tid due to copy&paste.

Now it's initialized in the test setup based on argv[0]. If argv[0] is a
path we take last element in it.

This patch is based on v1 by Li Wang but I've omitted moving the
parse_opts() up in the setup() since the setup() function expects
that tst_test sturcture is non-NULL and may segfault if that is not
true.

Signed-off-by: Li Wang <liwang@redhat.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Petr Vorel <pvorel@suse.cz>