ltp-aiodio: Create the file before fork
Currently, child's waiting for the file to appear in a loop with
usleep(100000) sometimes is too late to do the read before child
process is killed by parent process, which lead to child misses
the whole test.  Create and truncate the file before fork, then
drop the loop of waiting in read_sparse().

Signed-off-by: Guangwen Feng <fenggw-fnst@cn.fujitsu.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>