Revert "lib: Fix undefined reference to `mq_open' build failures"
This reverts commit fcfb6970002454cfa29cf679a82ba92c94fb59e0.

I've misunderstand how __attribute__((weak)) is supposed to work. The
problem here is that we link statically against the libltp library and
hence dynamic linker does not even try to resolve the library internal
weak mq_open symbol and we end up using the stub regardless if we linked
the library with -lrt or not.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>