syscalls/madvise09: Fix for disabled swap accounting
The test had to be fixed for systems with swap accounting disabled.

First there are no "memsw" prefixed tuning knobs in memcg directories if
swap accounting is disabled.

Then we had to limit the hungry child since otherwise it will be killed
only after all system swap was filled. Hence we exit the memory hungry
child once the process fills 100MB of swap. That seems to be enough to
put the cgroup into memory pressure while it limits the test runtime to
a few seconds.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>