commands/which01.sh: Fix test for debian.
The which binary on Debian comes from debianutils package and not from
the which package (like it is on the most of the distros). And since, as
far as I can tell, which command is not covered by POSIX or any other
standard, having slightly different version of the command is not a bug.

The debian version does not support most of the options, which works
fine, since the test produces TCONF for these but also returns relative
path instead of absolute one. So this patch adds alternate version for
the expected path and the test passes if at least one of them matches.

Fixes #151.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>